nameを分離
